test render (no asserts)

+++ tests/test_transformation/test_2d_matrix.cpp
@@ -23,7 +23,7 @@ TEST_CASE("test_transformation/test_2d_matrix | Same unit type", "[transformatio
 }
 
 TEST_CASE("test_transformation/test_2d_matrix | Apply transformation with scale", "[transformation/2d]") {
-    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::PixelsT, double>;
+    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::PixelsT, float>;
     TransformationType::Translation translation{10_px, 20_px};
     TransformationType::Scale scale = math::ratio(8_mm, 1_px);
     TransformationType::Rotation rotation{90.0_deg};
@@ -40,7 +40,7 @@ TEST_CASE("test_transformation/test_2d_matrix | Apply transformation with scale"
 }
 
 TEST_CASE("test_transformation/test_2d_matrix | Apply transformation without scale", "[transformation/2d]") {
-    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::MilimetersT, double>;
+    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::MilimetersT, float>;
     TransformationType::Translation translation{10_mm, 20_mm};
     TransformationType::Rotation rotation{90.0_deg};
 
@@ -56,7 +56,7 @@ TEST_CASE("test_transformation/test_2d_matrix | Apply transformation without sca
 }
 
 TEST_CASE("test_transformation/test_2d_matrix | Transform magnitude (with scale)", "[transformation/2d]") {
-    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::PixelsT, double>;
+    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::PixelsT, float>;
     TransformationType::Translation translation{10_px, 20_px};
     TransformationType::Scale scale = math::ratio(8_mm, 1_px);
     TransformationType::Rotation rotation{90.0_deg};
@@ -67,7 +67,7 @@ TEST_CASE("test_transformation/test_2d_matrix | Transform magnitude (with scale)
 }
 
 TEST_CASE("test_transformation/test_2d_matrix | Transform magnitude (without scale)", "[transformation/2d]") {
-    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::MilimetersT, double>;
+    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::MilimetersT, float>;
     TransformationType::Translation translation{10_mm, 20_mm};
     TransformationType::Rotation rotation{90.0_deg};
     auto m = TransformationType{translation, rotation};
@@ -79,7 +79,7 @@ TEST_CASE("test_transformation/test_2d_matrix | Transform magnitude (without sca
 
 
 TEST_CASE("test_transformation/test_2d_matrix | Set transformation elements (with scale)", "[transformation/2d]") {
-    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::PixelsT, double>;
+    using TransformationType = math::xy::types::Transformation<math::types::MilimetersT, math::types::PixelsT, float>;
     auto m = TransformationType{};
     auto inputPoint = Magnum::Math::Vector2{1_mm, 0_mm};
 
@@ -112,7 +112,7 @@ TEST_CASE("test_transformation/test_2d_matrix | Set transformation elements (wit
 }
 
 TEST_CASE("test_transformation/test_2d_matrix | Set transformation elements (without scale)", "[transformation/2d]") {
-    using TransformationType = math::xy::types::Transformation<math::types::PixelsT, math::types::PixelsT, double>;
+    using TransformationType = math::xy::types::Transformation<math::types::PixelsT, math::types::PixelsT, float>;
     auto m = TransformationType{};
     auto inputPoint = Magnum::Math::Vector2{1_px, 0_px};
