+++ tests/test_units/test_pixels.cpp
@@ -5,23 +5,23 @@
 using namespace math::units;
 
 TEST_CASE("test_units/test_pixels | Test constructor and literals (unsigned long long)", "[units/pixels]") {
-    math::types::PixelsT<double> pixels{3000};
+    math::types::PixelsT<float> pixels{3000};
     REQUIRE(pixels == 3000_px);
 }
 
 TEST_CASE("test_units/test_pixels | Test constructor and literals (long double)", "[units/pixels]") {
-    math::types::PixelsT<double> pixels{0.3};
+    math::types::PixelsT<float> pixels{0.3};
     REQUIRE(pixels == 0.3_px);
 }
 
 TEST_CASE("test_units/test_pixels | Test copy constructor", "[units/pixels]") {
-    math::types::PixelsT<double> mm1{0.3};
+    math::types::PixelsT<float> mm1{0.3};
     auto other = mm1;
     REQUIRE(mm1 == other);
 }
 
 TEST_CASE("test_units/test_pixels | Test output stream", "[units/pixels]") {
-    math::types::PixelsT<double> mm1{0.3};
+    math::types::PixelsT<float> mm1{0.3};
     std::ostringstream os;
     os << mm1;
     REQUIRE(os.str() == "0.3 px");
